requirements.txt
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.4
gunicorn==19.2.0
oauthlib==0.7.2
psycopg2==2.5.4
PyJWT==0.4.1
pystache==0.5.4
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
static==1.1.1
static3==0.5.1
